Split up views source file
